+++ R/plot.n_factors.R
@@ -3,43 +3,47 @@ data_plot.n_factors <- function(x, data = NULL, type = "bar", ...) {
   s1 <- summary(x)
 
   if ("n_Factors" %in% names(x)) {
-    var <- "n_Factors"
+    variable <- "n_Factors"
     lab <- "factors"
   } else {
-    var <- "n_Clusters"
+    variable <- "n_Clusters"
     lab <- "clusters"
   }
 
-  s2 <- data.frame(n_Methods = rep(0, max(x[[var]])))
+  s2 <- data.frame(n_Methods = rep(0, max(x[[variable]])))
 
   if (type == "line") {
-    s1[[var]] <- as.factor(s1[[var]])
-    s2[[var]] <- factor(1:max(x[[var]]))
+    s1[[variable]] <- as.factor(s1[[variable]])
+    s2[[variable]] <- factor(1:max(x[[variable]]))
   } else {
-    s2[[var]] <- 1:max(x[[var]])
+    s2[[variable]] <- 1:max(x[[variable]])
   }
 
-  if("Variance_Cumulative" %in% names(s1)){
+  if ("Variance_Cumulative" %in% names(s1)) {
     s2$Variance_Cumulative <- NA
   }
 
-  dataplot <- rbind(s1, s2[!s2[[var]] %in% s1[[var]], ])
+  dataplot <- rbind(s1, s2[!s2[[variable]] %in% s1[[variable]], ])
 
   # Add Variance explained
-  if("Variance_Explained" %in% names(attributes(x))){
+  if ("Variance_Explained" %in% names(attributes(x))) {
     dataplot$Variance_Cumulative <- NULL  # Remove column and re add
-    dataplot <- merge(dataplot, attributes(x)$Variance_Explained[, c("n_Factors", "Variance_Cumulative")], by = "n_Factors")
+    dataplot <- merge(
+      dataplot,
+      attributes(x)$Variance_Explained[, c("n_Factors", "Variance_Cumulative")],
+      by = "n_Factors"
+    )
   }
 
   if (type == "line") {
-    dataplot$x <- factor(dataplot[[var]], levels = rev(sort(levels(dataplot[[var]]))))
+    dataplot$x <- factor(dataplot[[variable]], levels = rev(sort(levels(dataplot[[variable]]))))
     dataplot$group <- "0"
     dataplot$group[which.max(dataplot$n_Methods)] <- "1"
   } else if (type == "area") {
-    dataplot$x <- dataplot[[var]]
+    dataplot$x <- dataplot[[variable]]
   } else {
-    dataplot <- dataplot[order(dataplot[[var]]), ]
-    dataplot$x <- dataplot[[var]]
+    dataplot <- dataplot[order(dataplot[[variable]]), ]
+    dataplot$x <- dataplot[[variable]]
     dataplot$fill <- "Not-optimal"
     dataplot$fill[which.max(dataplot$n_Methods)] <- "Optimal"
   }
@@ -54,23 +58,24 @@ data_plot.n_factors <- function(x, data = NULL, type = "bar", ...) {
   # Inverse xlab and ylab for line plot
   if (type == "line") {
     attr(dataplot, "info") <- list(
-      "ylab" = paste("Number of", lab),
-      "xlab" = axis_lab
+      ylab = paste("Number of", lab),
+      xlab = axis_lab
     )
   } else {
     attr(dataplot, "info") <- list(
-      "xlab" = paste("Number of", lab),
-      "ylab" = axis_lab
+      xlab = paste("Number of", lab),
+      ylab = axis_lab
     )
   }
   # Title
 
   attr(dataplot, "info")$title <-  paste("How many", lab, "to retain")
   attr(dataplot, "info")$subtitle <- paste0("Number of ", lab, " considered optimal by various algorithm")
-  if("Variance_Cumulative" %in% names(dataplot) && type != "line"){
+  if ("Variance_Cumulative" %in% names(dataplot) && type != "line") {
     attr(dataplot, "info")$subtitle <- paste0(
       attr(dataplot, "info")$subtitle,
-      ". The dashed line represent the cumulative percentage of variance explained")
+      ". The dashed line represent the cumulative percentage of variance explained"
+    )
   }
 
   class(dataplot) <- unique(c("data_plot", "see_n_factors", class(dataplot)))
@@ -124,8 +129,8 @@ plot.see_n_factors <- function(x,
 
   if (missing(size)) {
     size <- switch(type,
-      "bar" = 0.7,
-      "line" = 1,
+      bar = 0.7,
+      line = 1,
       1
     )
   }
@@ -168,11 +173,18 @@ plot.see_n_factors <- function(x,
   }
 
   # Add variance explained
-  if("Variance_Cumulative" %in% names(x)) {
+  if ("Variance_Cumulative" %in% names(x)) {
     x$Varex_scaled <- x$Variance_Cumulative * max(x$y)
     p <- p +
-      geom_line(data=x, aes(x = .data$x, y = .data$Varex_scaled, group=1), linetype="dashed") +
-      scale_y_continuous(labels = .percents, sec.axis = sec_axis(~ . / max(x$y), name = "% of variance explained", labels = .percents))
+      geom_line(
+        data = x,
+        aes(x = .data$x, y = .data$Varex_scaled, group = 1),
+        linetype = "dashed"
+      ) +
+      scale_y_continuous(
+        labels = .percents,
+        sec.axis = sec_axis(~ . / max(x$y), name = "% of variance explained", labels = .percents)
+      )
   } else {
     p <- p + scale_y_continuous(labels = .percents)
   }