+++ NEWS.md
@@ -1,3 +1,10 @@
+# see 0.8.2
+
+## Minor Changes
+
+* `plot.n_factors()` now shows a dashed line over the bars, indicating the
+  cumulate explained variance by the number of factors.
+
 # see 0.8.1
 
 ## Major Changes